requirements.txt
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
pkg-resources==0.0.0
python-dateutil==2.8.2
pytz==2021.1
scipy==1.5.4
six==1.16.0
SpeechRecognition==3.8.1
typing-extensions==3.10.0.2
wavio==0.0.4
Werkzeug==2.0.1
zipp==3.5.0
